requirements.txt
telethon==1.28.5
python-dotenv==1.0.0
loguru==0.7.0 